requirements.txt
streamlit
google-generativeai
folium
streamlit-folium
watchdog